[autoparallel] refactor handlers which reshape input tensors
